Simplify download and fix types
