+++ src/apps/main/network/download.types.ts
@@ -1,0 +1,14 @@
+import { DownloadProgressCallback } from '@internxt/inxt-js/build/lib/core';
+import { NetworkCredentials } from './requests';
+
+export interface IDownloadParams {
+  networkApiUrl: string;
+  bucketId: string;
+  fileId: string;
+  creds: NetworkCredentials;
+  mnemonic: string;
+  options: {
+    abortController?: AbortController;
+    notifyProgress: DownloadProgressCallback;
+  };
+}
