+++ src/apps/main/remote-sync/folders/sync-remote-folders.service.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-await-in-loop */
 import { RemoteSyncedFolder } from '../helpers';
 import { RemoteSyncManager } from '../RemoteSyncManager';
 import { FetchRemoteFoldersService } from './fetch-remote-folders.service';
