+++ src/context/local/localFile/application/update/FileBatchUpdater.ts
@@ -15,7 +15,6 @@ export class FileBatchUpdater {
 
   async run(localRoot: LocalFolder, remoteTree: RemoteTree, batch: Array<LocalFile>, signal: AbortSignal): Promise<void> {
     for (const localFile of batch) {
-      // eslint-disable-next-line no-await-in-loop
       const upload = await this.uploader.upload(localFile.path, localFile.size, signal);
 
       if (upload.isLeft()) {
@@ -32,7 +31,6 @@ export class FileBatchUpdater {
         throw new Error(`Expected file, found folder on ${file.path}`);
       }
 
-      // eslint-disable-next-line no-await-in-loop
       await this.simpleFileOverrider.run(file, contentsId, localFile.size);
     }
   }
